% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/about.R
\name{about}
\alias{about}
\title{Get a summary of OM properties}
\usage{
about(
  obj,
  xlim = c(-135, -125.5),
  ylim = c(48, 55.5),
  xfac = 1.7,
  col = "red",
  plot.fig = T,
  ret.dat = F,
  p = c(0.05, 0.95),
  plot.MAs = T
)
}
\arguments{
\item{obj}{An object of class In or OM or RCMfit}

\item{xlim}{Vector c(min, max) of longitudinal range to plot}

\item{ylim}{Vector c(min, max) of latitudinal range to plot}

\item{xfac}{Positive numerical value (probably above 1) that stretches the white space on the x axis around the plot}

\item{col}{Color of shaded area}

\item{plot.fig}{Logical Should the figure be plotted>}

\item{ret.dat}{Logical Should the summary data be returned?}

\item{p}{The quantiles to be calculated e.g. p=c(0.005, 0.995) is the 99 percent interval}

\item{plot.MAs}{Logical, should the management areas (statistical areas) be plotted (slower)?}
}
\description{
A plot and optional list of properties
}
\examples{
about(OM.GD.7)
}
\author{
T. Carruthers
}
