% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avail_objects.R
\name{objs}
\alias{objs}
\title{What objects of this class are available}
\usage{
objs(classy, package = NULL, msg = TRUE)
}
\arguments{
\item{classy}{A class of object (character string, e.g. 'In', 'OM', 'MSE', 'MP')}

\item{package}{Optional. Names(s) of the package to search for object of class `classy`. String
Default is all `Inverts` packages (Inverts, Inverts.GD, Inverts.MC, Inverts.GSU, Inverts.SC). Always searches the global environment as well.}

\item{msg}{Print messages?}
}
\description{
Generic class finder Finds objects of the specified class in the global environment or the
Inverts packages.
}
\examples{
objs("In", msg=FALSE)
RCM_input_files <- objs("In")
OMs <- objs("OM")
}
\seealso{
\link{avail}
}
\author{
T. Carruthers
}
