% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OM_mod.R
\name{OM_mod}
\alias{OM_mod}
\title{Modify natural mortality rate and recruitment}
\usage{
OM_mod(OM, M_val = NaN, M_trend = 0, R_trend = 0)
}
\arguments{
\item{OM}{An object of class OM}

\item{M_val}{A single value of M, an lower/upper range of M (uniformly sampled) or a per-simulation specification of M (vector nsim long)}

\item{M_trend}{A single value that is the annual percentage change in M (a value of 1 is a 1 percent increase per year in all years) or a custom pattern in M for each year that is a factor (1 is no change)}

\item{R_trend}{A single value that is the annual percentage change in recruitment strength (a value of 1 is a 1 percent increase per year in all years) or a custom pattern in M for each year that is a factor (1 is no change)}
}
\description{
A wrapper function that modifies the cpars slot of the OM object to control the value and trajectory of natural mortality rate and recruitment strength for all years
}
\examples{
OM_mod(OM.MC.E, M_val = c(0.2,0.3), M_trend = 1, Rtrend = -1)
}
\author{
T. Carruthers
}
