% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ref_pts.R
\name{Ref.Points}
\alias{Ref.Points}
\title{Summarize of biomass and F reference points}
\usage{
Ref.Points(
  Hist,
  plot = T,
  p = c(0.05, 0.25, 0.5, 0.75, 0.95),
  B0ref = c(0.2, 0.4),
  BMSYref = c(0.4, 0.8)
)
}
\arguments{
\item{Hist}{An object of class Hist created by runMSE(OM, Hist=T)}

\item{plot}{Logical, should the results be plotted?}

\item{p}{Vector of quantiles to plot (Default is 90%, 50% intervals plus median)}

\item{B0ref}{The SSB0 (unfished spawning biomass) reference levels (defaults to 0.2 and 0.4)}

\item{BMSYref}{The SSBMSY reference levels (defaults to 0.5 and 1.0)}
}
\description{
Tables and figures for visualizing historical biomass and exploitation rate reference points
}
\examples{
Ref.Points(Hist)
}
\author{
T. Carruthers
}
