% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_spawn.R
\name{spawnMP}
\alias{spawnMP}
\title{Spawns multiple MPs from a single MP}
\usage{
spawnMP(MPname, spawnlist, spawnnames = NA, defaults = list(NA))
}
\arguments{
\item{MPname}{The name of a function of class MP}

\item{spawnlist}{A list of argument names, each list are argument values nMPs long}

\item{spawnnames}{A character vector of MP names}

\item{defaults}{A list of default values for other arguments #' @author T. Carruthers}
}
\description{
An 'illegal' R function that allows you to rapidly make a large array of MPs
}
